refactor: remove changelog pkg
